requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.4.0
fpdf2>=2.7.9
openpyxl>=3.1.2
xlrd>=2.0.1